Pr feedback (#1)
* small fixes

* sensor check for uid duplicates

* entity_registry_enabled_default now poperty of UpdateCheckSensor

* Coordinator Separation

* Options schema validation is now enforced

* create entity factory

* sett permiision to actions

* fix unit tests
