+++ .github/workflows/ci.yml
@@ -23,9 +23,9 @@ jobs:
       - name: Check out code from GitHub
         uses: "actions/checkout@v4"
       - name: Black Code Format Check
-        uses: lgeiger/black-action@master
+        uses: psf/black@stable
         with:
-          args: ". --check --fast --diff"
+          options: ". --check --fast --diff"
 
   tests:
     name: Python Tests
@@ -64,12 +64,11 @@ jobs:
     steps:
       - name: Check out code from GitHub
         uses: "actions/checkout@v4"
+      - name: Install Bandit
+        run: pip install bandit
       - name: Security check - Bandit
-        uses: ioggstream/bandit-report-artifacts@v1.7.4
-        with:
-          project_path: custom_components/portainer
-          config_file: .github/bandit.yaml
-          ignore_failure: false
+        run: |
+          bandit -r custom_components/portainer -c .github/bandit.yaml -f txt -o output/security_report.txt
       - name: Security check report artifacts
         uses: actions/upload-artifact@v4
         with:
