+++ .github/workflows/stale.yml
@@ -1,8 +1,13 @@
-name: 'Stale'
+name: "Stale"
 on:
   schedule:
-    - cron: '30 1 * * *'
+    - cron: "30 1 * * *"
   workflow_dispatch:
+
+permissions:
+  contents: read
+  issues: write
+  pull-requests: write
 
 jobs:
   stale:
@@ -15,10 +20,10 @@ jobs:
     steps:
       - uses: "actions/stale@v9"
         with:
-          stale-issue-message: 'This issue has been automatically marked as stale because it has not had recent activity. It will be closed if no further activity occurs.'
-          stale-pr-message: 'This PR has been automatically marked as stale because it has not had recent activity. It will be closed if no further activity occurs.'
-          close-issue-message: 'This issue was closed because it has been stalled for 5 days with no activity.'
+          stale-issue-message: "This issue has been automatically marked as stale because it has not had recent activity. It will be closed if no further activity occurs."
+          stale-pr-message: "This PR has been automatically marked as stale because it has not had recent activity. It will be closed if no further activity occurs."
+          close-issue-message: "This issue was closed because it has been stalled for 5 days with no activity."
           days-before-stale: 14
           days-before-close: 7
-          exempt-issue-labels: 'pinned,security,planned,help wanted'
-          exempt-pr-labels: 'pinned,security,planned,help wanted'
+          exempt-issue-labels: "pinned,security,planned,help wanted"
+          exempt-pr-labels: "pinned,security,planned,help wanted"
