+++ custom_components/portainer/entity.py
@@ -19,206 +19,16 @@ from .const import (
     DOMAIN,
 )
 from .coordinator import PortainerCoordinator
+from .entity_factory import EntityFactory
 from .helper import format_attribute, format_camel_case
 
 _LOGGER = getLogger(__name__)
 
 
-# ---------------------------
-#   async_create_sensors
-# ---------------------------
-def _should_create_entity(description, data):
-    """Determine if an entity should be created based on description and data."""
-    if description.func == "UpdateCheckSensor":
-        return True
-    if (
-        data.get(description.data_attribute) is None
-        and description.func != "TimestampSensor"
-    ):
-        return False
-    return True
-
-
-def _create_temp_entity(dispatcher, func, coordinator, description, uid=None):
-    """Create a temporary entity object using the dispatcher."""
-    if uid is not None:
-        return dispatcher[func](coordinator, description, uid)
-    return dispatcher[func](coordinator, description)
-
-
-def _validate_entity(temp_obj, description, uid=None):
-    """Validate the temporary entity object for required properties."""
-    try:
-        unique_id = temp_obj.unique_id
-        entity_name = temp_obj.name
-    except (AttributeError, TypeError, KeyError) as e:
-        if uid is not None:
-            _LOGGER.error(
-                "Error accessing properties of entity %s (uid: %s): %s",
-                description.key,
-                uid,
-                e,
-            )
-        else:
-            _LOGGER.error(
-                "Error accessing properties of entity %s: %s",
-                description.key,
-                e,
-            )
-        return None, None
-    return unique_id, entity_name
-
-
-def _is_valid_entity(unique_id, entity_name, description, uid=None):
-    """Check if the entity has valid unique_id and name."""
-    if not unique_id or unique_id.strip() == "":
-        if uid is not None:
-            _LOGGER.warning(
-                "Skipping entity creation for %s (uid: %s): unique_id is None or empty (%s)",
-                description.key,
-                uid,
-                repr(unique_id),
-            )
-        else:
-            _LOGGER.warning(
-                "Skipping entity creation for %s: unique_id is None or empty (%s)",
-                description.key,
-                repr(unique_id),
-            )
-        return False
-    if not entity_name or entity_name.strip() == "":
-        if uid is not None:
-            _LOGGER.warning(
-                "Skipping entity creation for %s (uid: %s): name is None or empty (%s)",
-                description.key,
-                uid,
-                repr(entity_name),
-            )
-        else:
-            _LOGGER.warning(
-                "Skipping entity creation for %s: name is None or empty (%s)",
-                description.key,
-                repr(entity_name),
-            )
-        return False
-    return True
-
-
-def _final_entity_validation(entity):
-    """Final validation for entity before returning."""
-    try:
-        unique_id = entity.unique_id
-        entity_name = entity.name
-        entity_id = getattr(entity, "entity_id", None)
-
-        if not unique_id or unique_id.strip() == "":
-            _LOGGER.error(
-                "Filtering out entity with invalid unique_id: %s (name: %s, entity_id: %s)",
-                repr(unique_id),
-                repr(entity_name),
-                repr(entity_id),
-            )
-            return False
-        if not entity_name or entity_name.strip() == "":
-            _LOGGER.error(
-                "Filtering out entity with invalid name: %s (unique_id: %s, entity_id: %s)",
-                repr(entity_name),
-                repr(unique_id),
-                repr(entity_id),
-            )
-            return False
-
-        _LOGGER.debug(
-            "Final entity validation passed: unique_id=%s, name=%s, entity_id=%s",
-            unique_id,
-            entity_name,
-            entity_id,
-        )
-        return True
-    except (AttributeError, TypeError, KeyError) as e:
-        _LOGGER.error("Error validating entity during final check: %s", e)
-        return False
-
-
-def _add_entity_if_valid(new_entities, temp_obj, description, uid=None):
-    """Helper to validate and add entity if valid and not duplicate."""
-    unique_id, entity_name = _validate_entity(temp_obj, description, uid)
-    if not _is_valid_entity(unique_id, entity_name, description, uid):
-        return
-    if any(e.unique_id == unique_id for e in new_entities):
-        _LOGGER.debug(
-            "Entity with unique_id %s already in new_entities, skipping",
-            unique_id,
-        )
-        return
-    if uid is not None:
-        _LOGGER.debug(
-            "Adding entity with uid to new_entities: unique_id=%s, name=%s, uid=%s, type=%s",
-            unique_id,
-            entity_name,
-            uid,
-            type(temp_obj).__name__,
-        )
-    else:
-        _LOGGER.debug(
-            "Adding entity to new_entities: unique_id=%s, name=%s, type=%s",
-            unique_id,
-            entity_name,
-            type(temp_obj).__name__,
-        )
-    new_entities.append(temp_obj)
-
-
-def _process_description_without_reference(
-    new_entities, dispatcher, coordinator, description, data
-):
-    """Process a description without data_reference."""
-    if not _should_create_entity(description, data):
-        return
-    temp_obj = _create_temp_entity(
-        dispatcher, description.func, coordinator, description
-    )
-    _add_entity_if_valid(new_entities, temp_obj, description)
-
-
-def _process_description_with_reference(
-    new_entities, dispatcher, coordinator, description, data
-):
-    """Process a description with data_reference."""
-    for uid in data:
-        temp_obj = _create_temp_entity(
-            dispatcher, description.func, coordinator, description, uid
-        )
-        _add_entity_if_valid(new_entities, temp_obj, description, uid)
-
-
-def create_sensors(
-    coordinator: PortainerCoordinator, descriptions: list, dispatcher: dict
-) -> list[PortainerEntity]:
-    """Create Portainer sensor entities."""
-
-    new_entities = []
-
-    for description in descriptions:
-        if description.data_path not in coordinator.data:
-            coordinator.data[description.data_path] = {}
-
-        data = coordinator.data[description.data_path]
-        if not description.data_reference:
-            _process_description_without_reference(
-                new_entities, dispatcher, coordinator, description, data
-            )
-        else:
-            _process_description_with_reference(
-                new_entities, dispatcher, coordinator, description, data
-            )
-
-    final_entities = [
-        entity for entity in new_entities if _final_entity_validation(entity)
-    ]
-
-    _LOGGER.debug("Returning %d validated entities", len(final_entities))
-    return final_entities
+def create_sensors(coordinator, descriptions, dispatcher):
+    """Create Portainer sensor entities using EntityFactory."""
+    factory = EntityFactory(coordinator, dispatcher)
+    return factory.create_sensors(descriptions)
 
 
 # ---------------------------
