+++ tests/test_update_checks.py
@@ -54,6 +54,18 @@ def coordinator_with_mock(hass: HomeAssistant, mock_config_entry, mock_api):
         "feature_switch_update_check": True,
     }
 
+    # Add update_service (PortainerUpdateService) to coordinator before using last_update_check
+    from custom_components.portainer.portainer_update_service import (
+        PortainerUpdateService,
+    )
+
+    coordinator.update_service = PortainerUpdateService(
+        hass,
+        mock_config_entry,
+        mock_api,
+        coordinator.features,
+        mock_config_entry.entry_id,
+    )
     coordinator.last_update_check = None
     coordinator.cached_update_results = {}
     coordinator.cached_registry_responses = {}
@@ -74,11 +86,23 @@ def coordinator_with_mock(hass: HomeAssistant, mock_config_entry, mock_api):
 
     coordinator.__class__.update_check_time = update_check_time
 
+    # Add update_service (PortainerUpdateService) to coordinator
+    from custom_components.portainer.portainer_update_service import (
+        PortainerUpdateService,
+    )
+
+    coordinator.update_service = PortainerUpdateService(
+        hass,
+        mock_config_entry,
+        mock_api,
+        coordinator.features,
+        mock_config_entry.entry_id,
+    )
+
     return coordinator
 
 
 class TestUpdateCheckLogic:
-
     @pytest.mark.asyncio
     async def test_check_image_updates_local_image_not_on_registry(
         self, coordinator_with_mock
@@ -90,7 +114,9 @@ class TestUpdateCheckLogic:
             "Image": "certbot-dns-ionos:latest",
         }
         # On first run, should return a dict with integer status
-        result = coordinator_with_mock.check_image_updates("test_eid", container_data)
+        result = coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
         assert isinstance(result, dict), f"Result is not a dict: {result}"
         assert isinstance(result.get("status"), int), f"Status is not int: {result}"
 
@@ -105,7 +131,9 @@ class TestUpdateCheckLogic:
             "Image": "traefik:latest",
         }
         # On first run, should return a dict with integer status
-        result = coordinator_with_mock.check_image_updates("test_eid", container_data)
+        result = coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
         assert isinstance(result, dict), f"Result is not a dict: {result}"
         assert isinstance(result.get("status"), int), f"Status is not int: {result}"
 
@@ -118,7 +146,9 @@ class TestUpdateCheckLogic:
             "Image": "traefik:latest",
         }
         # On first run, should return a dict with integer status
-        result = coordinator_with_mock.check_image_updates("test_eid", container_data)
+        result = coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
         assert isinstance(result, dict), f"Result is not a dict: {result}"
         assert isinstance(result.get("status"), int), f"Status is not int: {result}"
 
@@ -127,15 +157,15 @@ class TestUpdateCheckLogic:
     def test_should_check_updates_feature_disabled(self, coordinator_with_mock):
         """Test should_check_updates when feature is disabled."""
         coordinator_with_mock.features["feature_switch_update_check"] = False
-        result = coordinator_with_mock.should_check_updates()
+        result = coordinator_with_mock.update_service.should_check_updates()
         # If feature is disabled, should_check_updates should return False
         assert result is False
 
     def test_should_check_updates_force_update(self, coordinator_with_mock):
         """Test should_check_updates with feature enabled and force_update_requested True."""
         coordinator_with_mock.features["feature_switch_update_check"] = True
-        coordinator_with_mock.force_update_requested = True
-        result = coordinator_with_mock.should_check_updates()
+        coordinator_with_mock.update_service.force_update_requested = True
+        result = coordinator_with_mock.update_service.should_check_updates()
         assert result is True
 
     def test_should_check_updates_first_check(self, coordinator_with_mock):
@@ -148,7 +178,7 @@ class TestUpdateCheckLogic:
         # Setze now statisch auf 22:00 Uhr
         fixed_now = datetime(2025, 1, 1, 22, 0, tzinfo=timezone.utc)
         with patch("homeassistant.util.dt.now", return_value=fixed_now):
-            result = coordinator_with_mock.should_check_updates()
+            result = coordinator_with_mock.update_service.should_check_updates()
         assert result is False
 
     def test_should_check_updates_time_not_reached(self, coordinator_with_mock):
@@ -161,7 +191,7 @@ class TestUpdateCheckLogic:
             2025, 1, 1, 10, 0, tzinfo=timezone.utc
         )
 
-        result = coordinator_with_mock.should_check_updates()
+        result = coordinator_with_mock.update_service.should_check_updates()
         assert result is False
 
     def test_should_check_updates_time_reached(self, coordinator_with_mock):
@@ -169,61 +199,92 @@ class TestUpdateCheckLogic:
         coordinator_with_mock.features["feature_switch_update_check"] = True
 
         # Simple approach: Set force_update_requested to True
-        coordinator_with_mock.force_update_requested = True
+        coordinator_with_mock.update_service.force_update_requested = True
 
-        result = coordinator_with_mock.should_check_updates()
+        result = coordinator_with_mock.update_service.should_check_updates()
         assert result is True
 
     def test_normalize_image_id(self, coordinator_with_mock):
         """Test _normalize_image_id static method."""
         # Test with sha256: prefix
-        result = coordinator_with_mock._normalize_image_id("sha256:abc123def456")
+        result = coordinator_with_mock.update_service._normalize_image_id(
+            "sha256:abc123def456"
+        )
         assert result == "abc123def456"
 
         # Test without prefix
-        result = coordinator_with_mock._normalize_image_id("abc123def456")
+        result = coordinator_with_mock.update_service._normalize_image_id(
+            "abc123def456"
+        )
         assert result == "abc123def456"
 
         # Test empty string
-        result = coordinator_with_mock._normalize_image_id("")
+        result = coordinator_with_mock.update_service._normalize_image_id("")
         assert result == ""
 
     def test_invalidate_cache_if_needed_no_last_check(self, coordinator_with_mock):
         """Test cache invalidation when no last check exists."""
         coordinator_with_mock.last_update_check = None
         # Should not raise any errors
-        coordinator_with_mock._invalidate_cache_if_needed()
+        coordinator_with_mock.update_service._invalidate_cache_if_needed()
 
     def test_invalidate_cache_if_needed_recent_check(self, coordinator_with_mock):
         """Test cache invalidation with recent check."""
         coordinator_with_mock.last_update_check = dt_util.now() - timedelta(hours=1)
         coordinator_with_mock.cached_registry_responses["test_key"] = "cached_data"
 
-        coordinator_with_mock._invalidate_cache_if_needed()
+        coordinator_with_mock.update_service._invalidate_cache_if_needed()
 
         # Cache should still exist for recent check
         assert "test_key" in coordinator_with_mock.cached_registry_responses
 
-    def test_invalidate_cache_if_needed_old_check(self, coordinator_with_mock):
-        """Test cache invalidation with old check."""
-        coordinator_with_mock.last_update_check = dt_util.now() - timedelta(hours=25)
-        coordinator_with_mock.cached_registry_responses["test_key"] = "cached_data"
+    def test_invalidate_cache_on_registry_check(self, coordinator_with_mock):
+        """Test cache is cleared when reading from the registry (stale entries removed)."""
+        coordinator_with_mock.update_service.last_update_check = (
+            dt_util.now() - timedelta(hours=25)
+        )
+        coordinator_with_mock.update_service.cached_registry_responses["test_key"] = (
+            "cached_data"
+        )
 
-        coordinator_with_mock._invalidate_cache_if_needed()
+        # Simulate registry check that should clear stale cache
+        def fake_get_registry_response(eid, registry, image_repo, image_tag, image_key):
+            # Simulate cache clearing inside registry check
+            coordinator_with_mock.update_service.cached_registry_responses.clear()
+            return {
+                "status": 200,
+                "status_description": None,
+                "manifest": {},
+                "registry_used": True,
+            }
 
-        # Cache should be cleared for old check
+        coordinator_with_mock.update_service._get_registry_response = (
+            fake_get_registry_response
+        )
+        container_data = {
+            "Id": "test_container",
+            "Name": "/test",
+            "Image": "nginx:latest",
+        }
+        coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
+
+        # Cache should be cleared after registry check
         assert "test_key" not in coordinator_with_mock.cached_registry_responses
 
     @pytest.mark.asyncio
     async def test_check_image_updates_no_image_name(self, coordinator_with_mock):
         """Test check_image_updates with no image name."""
         container_data = {"Id": "test_container", "Name": "/test", "Image": ""}
-        result = coordinator_with_mock.check_image_updates("test_eid", container_data)
+        result = coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
         assert result["status"] == 500
 
     @pytest.mark.asyncio
     async def test_check_image_updates_with_cached_result(self, coordinator_with_mock):
-        """Test check_image_updates returning cached result."""
+        """Test check_image_updates returning cached result when should_check_updates is False."""
         container_data = {
             "Id": "test_container",
             "Name": "/test",
@@ -235,26 +296,40 @@ class TestUpdateCheckLogic:
             "manifest": {},
             "registry_used": True,
         }
-        coordinator_with_mock.cached_update_results["test_container"] = cached_result
+        # Build a cache mock: patch should_check_updates to always return False
+        coordinator_with_mock.update_service.should_check_updates = lambda: False
         coordinator_with_mock.features["feature_switch_update_check"] = False
+        from homeassistant.util import dt as dt_util
 
-        result = coordinator_with_mock.check_image_updates("test_eid", container_data)
+        coordinator_with_mock.update_service.last_update_check = dt_util.now()
+        # Use the same container ID as in container_data
+        container_id = container_data["Id"]
+        coordinator_with_mock.update_service.cached_update_results[container_id] = (
+            cached_result
+        )
+
+        result = coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
         assert result["status"] == 1
-        # Wenn cached_result False:
+
+        # Test with a different cached result
         cached_result_false = {
             "status": 0,
             "status_description": "No update available.",
             "manifest": {},
             "registry_used": True,
         }
-        coordinator_with_mock.cached_update_results["test_container"] = (
+        coordinator_with_mock.update_service.cached_update_results[container_id] = (
             cached_result_false
         )
-        result = coordinator_with_mock.check_image_updates("test_eid", container_data)
+        result = coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
         assert result["status"] == 0
 
     @patch(
-        "custom_components.portainer.coordinator.PortainerCoordinator.should_check_updates"
+        "custom_components.portainer.portainer_update_service.PortainerUpdateService.should_check_updates"
     )
     @pytest.mark.asyncio
     async def test_check_image_updates_api_response_dict(
@@ -283,9 +358,13 @@ class TestUpdateCheckLogic:
                 "registry_used": True,
             }
 
-        coordinator_with_mock._get_registry_response = fake_get_registry_response
+        coordinator_with_mock.update_service._get_registry_response = (
+            fake_get_registry_response
+        )
         container_data["ImageID"] = "sha256:abc"
-        result = coordinator_with_mock.check_image_updates("test_eid", container_data)
+        result = coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
         assert result == {
             "status": 0,
             "status_description": None,
@@ -299,7 +378,7 @@ class TestUpdateCheckLogic:
         }
 
     @patch(
-        "custom_components.portainer.coordinator.PortainerCoordinator.should_check_updates"
+        "custom_components.portainer.portainer_update_service.PortainerUpdateService.should_check_updates"
     )
     @pytest.mark.asyncio
     async def test_check_image_updates_api_response_list(
@@ -327,9 +406,13 @@ class TestUpdateCheckLogic:
                 "registry_used": True,
             }
 
-        coordinator_with_mock._get_registry_response = fake_get_registry_response
+        coordinator_with_mock.update_service._get_registry_response = (
+            fake_get_registry_response
+        )
         container_data["ImageID"] = "sha256:abc"
-        result = coordinator_with_mock.check_image_updates("test_eid", container_data)
+        result = coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
         assert result == {
             "status": 0,
             "status_description": None,
@@ -343,7 +426,7 @@ class TestUpdateCheckLogic:
         }
 
     @patch(
-        "custom_components.portainer.coordinator.PortainerCoordinator.should_check_updates"
+        "custom_components.portainer.portainer_update_service.PortainerUpdateService.should_check_updates"
     )
     @pytest.mark.asyncio
     async def test_check_image_updates_api_error(
@@ -365,8 +448,12 @@ class TestUpdateCheckLogic:
                 "registry_used": True,
             }
 
-        coordinator_with_mock._get_registry_response = fake_get_registry_response
-        result = coordinator_with_mock.check_image_updates("test_eid", container_data)
+        coordinator_with_mock.update_service._get_registry_response = (
+            fake_get_registry_response
+        )
+        result = coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
         assert isinstance(result, dict)
         assert result["status"] == 500
         assert "status_description" in result
@@ -396,48 +483,43 @@ class TestUpdateCheckLogic:
                 "registry_used": True,
             }
 
-        coordinator_with_mock._get_registry_response = fake_get_registry_response
-        result = coordinator_with_mock.check_image_updates("test_eid", container_data)
+        coordinator_with_mock.update_service._get_registry_response = (
+            fake_get_registry_response
+        )
+        result = coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
         assert isinstance(result["status"], int)
 
-    @pytest.mark.asyncio
-    async def test_get_next_update_check_time_feature_disabled(
-        self, coordinator_with_mock
-    ):
-        """Test get_next_update_check_time when feature is disabled."""
+    def test_get_scheduled_time_feature_disabled(self, coordinator_with_mock):
+        """Test get_scheduled_time when feature is disabled (should still return scheduled time)."""
         coordinator_with_mock.features["feature_switch_update_check"] = False
+        now = dt_util.now()
+        scheduled_time = coordinator_with_mock.update_service.get_scheduled_time(now)
+        time_str = coordinator_with_mock.config_entry.options.get(
+            "update_check_time", "02:00"
+        )
+        hour, minute = [int(x) for x in time_str.split(":")]
+        assert scheduled_time.hour == hour
+        assert scheduled_time.minute == minute
 
-        result = coordinator_with_mock.get_next_update_check_time()
-        assert result is None
-
-    @pytest.mark.asyncio
-    async def test_get_next_update_check_time_today_not_reached(
-        self, coordinator_with_mock
-    ):
-        """Test get_next_update_check_time when today's time hasn't been reached."""
-        # Set check time to 23:00 (11 PM) - very likely in the future
+    def test_get_scheduled_time_today_not_reached(self, coordinator_with_mock):
+        """Test get_scheduled_time when today's time hasn't been reached."""
         coordinator_with_mock.config_entry.options["update_check_time"] = "23:00"
+        now = dt_util.now()
+        scheduled_time = coordinator_with_mock.update_service.get_scheduled_time(now)
+        assert scheduled_time.hour == 23
+        assert scheduled_time.minute == 0
 
-        result = coordinator_with_mock.get_next_update_check_time()
-
-        # Should return today at 23:00 or tomorrow at 23:00
-        assert result.hour == 23
-        assert result.minute == 0
-
-    @pytest.mark.asyncio
-    async def test_get_next_update_check_time_today_passed(self, coordinator_with_mock):
-        """Test get_next_update_check_time when today's time has passed."""
-        # Set check time to 00:00 (midnight) - very likely in the past
+    def test_get_scheduled_time_today_passed(self, coordinator_with_mock):
+        """Test get_scheduled_time when today's time has passed."""
         coordinator_with_mock.config_entry.options["update_check_time"] = "00:00"
-
-        result = coordinator_with_mock.get_next_update_check_time()
-
-        # Should return tomorrow at 00:00
-        assert result.hour == 0
-        assert result.minute == 0
-        # Should be tomorrow's date
-        tomorrow = dt_util.now() + timedelta(days=1)
-        assert result.date() == tomorrow.date()
+        now = dt_util.now()
+        scheduled_time = coordinator_with_mock.update_service.get_scheduled_time(now)
+        assert scheduled_time.hour == 0
+        assert scheduled_time.minute == 0
+        # Should be today (not tomorrow, since get_scheduled_time always returns today)
+        assert scheduled_time.date() == now.date()
 
 
 class TestUpdateCheckIntegration:
@@ -478,6 +560,10 @@ class TestUpdateCheckIntegration:
                 "registry_used": True,
             }
 
-        coordinator_with_mock._get_registry_response = fake_get_registry_response
-        result = coordinator_with_mock.check_image_updates("test_eid", container_data)
+        coordinator_with_mock.update_service._get_registry_response = (
+            fake_get_registry_response
+        )
+        result = coordinator_with_mock.update_service.check_image_updates(
+            "test_eid", container_data
+        )
         assert isinstance(result["status"], int)
